能打开

+++ app/src/main/java/com/warrior/hangsu/administrator/strangerbookreader/bean/BookBean.java
@@ -45,7 +45,7 @@ public class BookBean {
         return progress;
     }
 
-    public void setProgress(int progress) {
+    public void setProgress(float progress) {
         this.progress = progress;
     }
